Make custom items spawn in hands before being dropped on the floor (#21417)
A small safety net. Currently when people latejoin, a custom item such
as a journal risk being dropped into the lift or cryo, where it can't be
retrived from without admin help, if their bags are filled from the
loadout. This fixes that issue by putting it into peoples hands first if
possible.
Also it makes the alert for when it is dropped on the ground be nice and
red, as otherwise it was way too easy to miss between all the other
messages from spawning in.