Adds Timers to Public Tool Storage (#21419)
Turns out- Timers have sat in the contraband section of the Vendomat in
tool storage all this time.

Previously, the only way to get them was through Autolathes or Grenade
Cases in the pharmacy and research chemistry lab. This makes the ones
already in the tool storage accessible. Cooks can have a kitchen timer,
or security can use it for target practice drills, or whatever you want,
really. It's a timer!

I doubt local grey-shirt self-antagging himself a single-tank-bomb is as
valid a worry as it was back when vending machines were originally being
added.

---------

Signed-off-by: HanSolo1519 <72716882+HanSolo1519@users.noreply.github.com>