Add journals and notepads with blank paper (#21418)
They are going to be filled anyway, and this way people won't have to
raid paper bins nearly as much. It adds a variant of the notepad and the
journal that is pre-filled with 5 blank papers. It also adds 5 blank
papers to my own custom item as an extension of this.
Also this fixes a runtime error when journals with content were deleted.