Drobe re-arrangement part two (#21416)
After a discussion a little while ago, it was agreed drobes should not
have storage objects in them aside from jackets, or anything that is
mechanically job related pretty much, just fluff and decorative stuff.
As a result, all bags, holsters, webbings and the janitor's galoshes
were removed from them as they belong in their respective lockers. This
does throw a bundle of things into certain lockers to permit selecting
it, such as the officers having various holsters and engineers having
the alt toolbelt.
Also hangar techs aren't getting their insulated gloves anymore, as that
was apparently a mistake.
Engineering lockers as a result of this do have insulated gloves now, as
it is relevant work equipment. No more raiding the electrical lockers.