Psionic Awakener (#21413)
This PR unhardcodes the psionic jumpstarter, allowing it to be utilized
by Skrell to upgrade their psionics to the level of an antagonist. To
make use of this new code, I've added a new Psionic Awakener, which is a
variant of the jumpstarter that sets the user to the level of a Skrell
psion. The awakener is significantly cheaper than a full jumpstarter,
and can be purchased using bluecrystals, which reflects its status as a
harmless item given that none of the Sensitive rank powers are overtly
harmful.