Fix - Issue & PR triage action (#21453)
# Summary

This PR fixes the to-be exempt labels on issues and PRs that the stale
action should ignore when trying to mark them as stale.

# Details

The label exempt regex syntax isn't supported [anymore?], this resulted
in the setting being ignored.
This PR also contains light cleaning, removal of defaults, update of
cron to daily to utilize the rate limit per run without getting rate
limited.

# On merge

Please trigger the action manually on merge so the output is immediately
available to verify the fix.
The documentation provides no info on the formatting of the labels
within their comma brackets, it might be possible they have another
format in code (e.g. `"Bot Ignore"` -> `"bot-ignore"`?).
Not a real breaking behavior if wrong but possibly annoying and open to
fix afterwards.