Adjust triage bot rate limit (#21471)
# Summary

This PR updates the triage bots schedule and raises the
operations-per-run limit.

# Details

Previous actions have been implemented using a defensive approach to
test the bots runtime and rate limits.

At the moment the single run at 6 in the morning utilized `0,0066`% of
our hourly GitHub API rate limit.
With those settings we will be done with the currently open 800 issues
in ~1/3 year.

This PR updates the schedule to hourly and raises the API limit to 100  
(3x of the default, out of 4500 API calls per hour, based of the stale
action execution log).

- Hourly will make the bot much more responsive (after the 800 issues
have been ~~purged~~ processed).
- A higher rate will allow the current state to be processed faster and
will provide better response to higher workloads in the future.
- The cache continues to handle any workload above 100
operations-per-run.