Bar Disposal Pipe Fix (#21503)
I accidentally double piped and used an improper junction and apparently
didn't catch it until now. This fixes that so the upper bar disposal
works. Obfuscation was automatically deleted so I'm not sure if that was
needed or not but it looks like it was only one.

The red square is where the extra pipe was. 

<img width="85" height="239" alt="image"
src="https://github.com/user-attachments/assets/2bb928ae-0848-4dc9-af09-2ebc52957340"
/>