Biogenerator Hint Tweaks (#21515)
Mechanics hints were slightly reworked in a separate PR from the
biogenerator to allow objects themselves to define the style and bullet
points of hints, but the biogenerator hints were updated to show dynamic
changes in a separate PR from that one. This PR just tweaks the hints
for the biogenerator to follow the same style as other upgradeable
machines such as the protolathe.