removes some mentions of Jargon in code. (#21553)
It hasnt been jargon for years, and the code still had a few references
I noticed. Most of these are just icon names, which arent at all visible
on the player side, but one description for specifically the cargo order
of a item still mentioned it. Funnily enough, the actual item itself had
a correct description.