We're Sol Back: The Solarian Frigate (#21078)
This is my first time mapping as a code-illiterate goober, so there's
probably a lot of silly mistakes in here.

Adds a very WIP build of a Solarian Frigate, meant to be deployed rarely
in sectors close to the borders of the ASSN (eg. the Badlands, Valley
Hale, and **_mayyyyyyyybe_** the Crescent Expanse. )

Against, most of this offship is very WIP: The code is close to
non-functional, the power grid just doesn't connect, and some areas just
look visually bland in my opinion. But the basic outline of the ship is
there.

<img width="8160" height="8160" alt="StrongDMM-2025-07-31 01 43 02"
src="https://github.com/user-attachments/assets/9b9a82fd-adcc-42f2-8ecb-d8ea6d43d39e"
/>


Overall, it's a ship that's just a bit smaller than the AXSF frigate
with a CIC-centric layout and equipped with a Longbow and Grauwolf
(pending replacement with the light coilgun, whenever that gets added.)
Honestly, it's just a regular corvette but scaled up. I'll add a second
Z-level or something else that might spice things up later.

---------

Signed-off-by: mynameasjeff <55303178+mynameasjeff@users.noreply.github.com>
Co-authored-by: Kano <89972582+kano-dot@users.noreply.github.com>
Co-authored-by: kano-dot <bhutanlikanoxy@gmail.com>