Warehouse rework (#21561)
The warehouse has been reworked to give it more floor space for the
crates, which now also start on the floor to make them easier to sort.
The floor was also changed there to make it feel more like a warehouse.
The upper warehouse has been changed to be catwalks leading to the break
room, mainly for the atmosphere, and the hangar tech lockers were moved
to a side part of the break room instead of free standing in the
hallway.