Nerfs suplexes. (#21572)
Suplexes now only deal 15 brute on hit, and now stun the attacker for as
much time as they stun the victim. This reframes the use of a suplex not
as a free kill tool that should be spammed at every individual, but
instead as a tool to reset a fight to a neutral state, including
auto-disarming the other party and equalising the playing field.

I have seen a significant number of rounds where the antagonist just
goes around suplex spamming everyone. It sucks.