Enforce all taj to start with siik'maas (#21605)
After talking with Catsin, it was agreed to enforce that all taj get
siik'maas as the edge cases of a taj not having it are few and very far
between.
It does not change their extra optional languages, letting them still
learn 2 extra for a total of 4 languages. The forced ceti basic and
siik'maas, and then the optional ones are between tradeband, freespeak,
sign, siik'tajr and one of the subspecies specific ones (yassa,
nal'arasan, delvahhi).