Laptop Vendor now dispenses more things (#21631)
My long-awaited return to coding.

The Laptop Vendor had a ton of stuff commented out, likely because it
was added before many of these parts were ported to Aurora. In addition,
it was unable to vend PDAs or tablets. It can now do so.

This machine is really in need of a re-do, since it's rather clunky. But
this was a quick option.

---------

Co-authored-by: GeneralCamo <n/a>