Fix kneepads on pants (#21660)
# Summary

This PR fixes kneepads not being able to be attached to pants.

- Fixes #21346

## Changes

- Change accessory slot type of kneepads to pants only.
- Added allowed pants accessory slot type to clothing/under.

## Notes

The preview for kneepads on pants in the char setup is not possible at
the moment. This is not related to the fix as it is another issue.

You can select the uniform for the kneepads in your loadout to let them
spawn in your backpack instead of having to remove them from your
jumpsuit or similar on spawn as a workaround to pants not being able to
be selected as an accessory target in the loadout.