Fixes the uniform hiding button in character setup (#21675)
The hide uniform, suit and hat buttons now work in character setup. I
changed it slightly so it will work regardless of it being a loadout or
job item in the specific slot, mainly because that was easier but also
because sometimes you want to change the uniform but the suit is in the
way.