Added a games vending machine (#21678)
It's unfortunate that the only way to get most of the toys is through
the loadout. This essentially restricts players to either roleplaying as
diehard fans, arranging games in advance, or not bothering to do this at
all.

This ports a games vending machine from bay, and places it in the
library. This supplements the battle monsters machine already present.

### Asset Licenses
The following assets that **have not** been created by myself are
included in this PR:

| Path | Original Author | License |
| --- | --- | --- |
| icons/obj/vending.dmi |
[nearlyNonexistent](https://github.com/Baystation12/Baystation12/commits?author=nearlyNonexistent)
(Baystation12) | CC-BY-SA 3.0 |