Fixes the colorable palette on colorable dresses (#21652)
> - bugfix: "Fixes the colorable palette used on colorable dresses."
  
Long story short, the dresses introduced in
https://github.com/Aurorastation/Aurora.3/pull/18644 were repaletted to
the new colorable clothing palette in
https://github.com/Aurorastation/Aurora.3/pull/21006. But that repalette
has had the colors put in the wrong order, resulting in a worse quality
sprite.
  
Issue presented on the longsleeve A-line dress, what's currently in game
on the left, and how it is supposed to look like on the right.

<img width="2048" height="1025" alt="the dresses"
src="https://github.com/user-attachments/assets/4d906ceb-fc42-4aee-aa8b-33fa39f60927"
/>